finished with basic login flow
