--- SpotifyGames/frontend/Components/UserLogin.js
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import SpotifyLogin from './SpotifyLoginScreen';
-import TestProfile from './TestProfile';
-
-export default function UserLogin(){
-
-
-    return (
-    <>
-        <SpotifyLogin/>
-        <TestProfile/>
-    </>
-    )
-}
